requirements.txt
Pillow==8.0.1
python-dotenv==0.9.1
requests==2.25.0
instabot==0.117.0